Anpassung von String zu float

+++ VS_Code_Version/src/automatic_irrigation.cpp
@@ -7,10 +7,10 @@
 void automaticIrrigation()
 {
     // Auslesen der Sensorwerte
-    float tankLevel = getTankLevel();      // Implementierung benötigt, aktuell nur Dummy-Wert
-    float temperature = readTemperature(); // Temperatur auslesen aus temp_sensor.cpp
-    float soilMoisture = readsoil_sensor().toFloat();
-    float humidity = readHumidity(); // Luftfeuchtigkeit auslesen aus temp_sensor.cpp
+    float tankLevel = getTankLevel();       // Implementierung benötigt, aktuell nur Dummy-Wert
+    float temperature = readTemperature();  // Temperatur auslesen aus temp_sensor.cpp
+    float soilMoisture = readsoil_sensor(); // Bodenfeuchtigkeit auslesen aus soil_sensor.cpp
+    float humidity = readHumidity();        // Luftfeuchtigkeit auslesen aus temp_sensor.cpp
 
     // Bedingungen für das Einschalten der Pumpe
     if (tankLevel > 5 && temperature < 35 && soilMoisture < 50)
